Fixed fonts and colors
